[fix]using path as ID in RTSP server
