+++ rtsp/rtsp-client.go
@@ -3,14 +3,11 @@ package rtsp
 import (
 	"bufio"
 	"bytes"
-	"crypto/md5"
-	"encoding/base64"
 	"encoding/binary"
 	"fmt"
 	"io"
 	"net"
 	"net/url"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -94,95 +91,6 @@ func NewRTSPClient(
 	}
 
 	return
-}
-
-// BasicAuth of RTSP
-func BasicAuth(authLine string, method string, URL string) (string, error) {
-	l, err := url.Parse(URL)
-	if err != nil {
-		return "", fmt.Errorf("Url parse error:%v,%v", URL, err)
-	}
-
-	username := l.User.Username()
-	password, _ := l.User.Password()
-	basic := base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s:%s", username, password)))
-
-	authorization := fmt.Sprintf("Basic %s", basic)
-
-	return authorization, nil
-}
-
-// DigestAuth of RTSP
-func DigestAuth(authLine string, method string, URL string) (string, error) {
-	l, err := url.Parse(URL)
-	if err != nil {
-		return "", fmt.Errorf("Url parse error:%v,%v", URL, err)
-	}
-	realm := ""
-	nonce := ""
-	realmRex := regexp.MustCompile(`realm="(.*?)"`)
-	result1 := realmRex.FindStringSubmatch(authLine)
-
-	nonceRex := regexp.MustCompile(`nonce="(.*?)"`)
-	result2 := nonceRex.FindStringSubmatch(authLine)
-
-	if len(result1) == 2 {
-		realm = result1[1]
-	} else {
-		return "", fmt.Errorf("auth error : no realm found")
-	}
-	if len(result2) == 2 {
-		nonce = result2[1]
-	} else {
-		return "", fmt.Errorf("auth error : no nonce found")
-	}
-	// response= md5(md5(username:realm:password):nonce:md5(public_method:url));
-	username := l.User.Username()
-	password, _ := l.User.Password()
-	l.User = nil
-	if l.Port() == "" {
-		l.Host = fmt.Sprintf("%s:%s", l.Host, "554")
-	}
-	md5UserRealmPwd := fmt.Sprintf("%x", md5.Sum([]byte(fmt.Sprintf("%s:%s:%s", username, realm, password))))
-	md5MethodURL := fmt.Sprintf("%x", md5.Sum([]byte(fmt.Sprintf("%s:%s", method, l.String()))))
-
-	response := fmt.Sprintf("%x", md5.Sum([]byte(fmt.Sprintf("%s:%s:%s", md5UserRealmPwd, nonce, md5MethodURL))))
-	Authorization := fmt.Sprintf("Digest username=\"%s\", realm=\"%s\", nonce=\"%s\", uri=\"%s\", response=\"%s\"", username, realm, nonce, l.String(), response)
-	return Authorization, nil
-}
-
-func (client *RTSPClient) checkAuth(method string, resp *Response) (string, error) {
-	if resp.StatusCode == 401 {
-
-		// need auth.
-		AuthHeaders := resp.Header["WWW-Authenticate"]
-
-		if auths, ok := AuthHeaders.([]string); ok {
-			for _, authLine := range auths {
-
-				if strings.IndexAny(authLine, "Digest") == 0 {
-					// 					realm="HipcamRealServer",
-					// nonce="3b27a446bfa49b0c48c3edb83139543d"
-					client.authLine = authLine
-					return DigestAuth(authLine, method, client.URL)
-				} else if strings.IndexAny(authLine, "Basic") == 0 {
-					client.authLine = authLine
-					return BasicAuth(authLine, method, client.URL)
-				}
-
-			}
-			return "", fmt.Errorf("auth method nnot support, %v", auths)
-		} else if authLine, ok := AuthHeaders.(string); ok {
-			if strings.IndexAny(authLine, "Digest") == 0 {
-				client.authLine = authLine
-				return DigestAuth(authLine, method, client.URL)
-			} else if strings.IndexAny(authLine, "Basic") == 0 {
-				client.authLine = authLine
-				return BasicAuth(authLine, method, client.URL)
-			}
-		}
-	}
-	return "", nil
 }
 
 func (client *RTSPClient) requestStream(timeout time.Duration) (err error) {
@@ -372,9 +280,10 @@ func (client *RTSPClient) requestStream(timeout time.Duration) (err error) {
 }
 
 func (client *RTSPClient) startStream() {
+	defer client.Stop()
+	
 	startTime := time.Now()
 	loggerTime := time.Now().Add(-10 * time.Second)
-	defer client.Stop()
 	for !client.Stoped {
 		if client.OptionIntervalMillis > 0 {
 			if time.Since(startTime) > time.Duration(client.OptionIntervalMillis)*time.Millisecond {
